wg-quick: move shell script bin into icedos toolset

+++ system/applications/modules/wg-quick/default.nix
@@ -20,19 +20,34 @@ mkIf (cfg.enable) {
     }) cfg.interfaces
   );
 
-  environment.systemPackages = [
-    (pkgs.writeShellScriptBin "wg-add-config-file" ''
-      sudo bash -c '
-        set -e
-        newFile="/etc/wireguard/''$(basename $1)"
+  icedos.internals.toolset.commands = [
+    (
+      let
+        command = "wg-config";
+      in
+      {
+        bin = "${pkgs.writeShellScript command ''
+          sudo bash -c '
+            set -e
 
-        mkdir -p /etc/wireguard
-        umask 0022 /etc/wireguard
+            if [[ "$1" == "" ]]; then
+              echo "error: provide config file location as an argument"
+              exit 1
+            fi
 
-        cp "$1" "$newFile"
-        chmod 600 "$newFile"
-        rm "$1"
-      ' -- $@
-    '')
+            newFile="/etc/wireguard/''$(basename $1)"
+
+            mkdir -p /etc/wireguard
+            umask 0022 /etc/wireguard
+
+            cp "$1" "$newFile"
+            chmod 600 "$newFile"
+            rm "$1"
+          ' -- $@
+        ''}";
+        command = command;
+        help = "add wireguard config to /etc/wireguard";
+      }
+    )
   ];
 }
